+++ url-updater/src/main/java/com/devonfw/tools/ide/url/folderhandling/jsonfile/JsonDataBlockForSpecificUrl.java
@@ -4,21 +4,21 @@ import java.util.Set;
 
 public class JsonDataBlockForSpecificUrl {
 
-	private Set<Double> urlHashes;
+	private Set<Integer> urlHashes;
 
 	public JsonDataBlockForSpecificUrl() {
 
 		super();
 	}
 
-	// TO DO: make one block per url, not per file with hashes for the urls in one
+	// TODO: Make one block per url, not per file with hashes for the urls in one
 	// list.
-	public JsonDataBlockForSpecificUrl(Set<Double> urlHashes) {
+	public JsonDataBlockForSpecificUrl(Set<Integer> urlHashes) {
 
 		this.urlHashes = urlHashes;
 	}
 
-	public Set<Double> getUrlHashes() {
+	public Set<Integer> getUrlHashes() {
 
 		return urlHashes;
 	}
